backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.0
gunicorn==21.2.0
google-generativeai==0.3.2
